requirements.txt
azure-core==1.30.0
azure-cosmos==4.5.1
azure-data-tables==12.5.0
azure-storage-blob==12.19.0
azure-storage-queue==12.9.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
dnspython==2.6.1
Flask==2.2.2
gunicorn==21.2.0
idna==3.6
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
packaging==23.2
pip==24.0
pycparser==2.21
pymongo==4.6.2
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
typing-extensions==4.10.0
urllib3==2.2.1
Werkzeug==2.2.2
yarl==1.9.4
